backend/requirements.txt
requests==2.31.0
feedparser==6.0.10
schedule==1.2.1
python-dotenv==1.0.1
pyyaml==6.0.1
